+++ tg_sdk/customer/customer.py
@@ -1,5 +1,0 @@
-from tg_sdk.abstract import RetrieveResourceMixin
-
-
-class Customer(RetrieveResourceMixin):
-    resource = 'customers'