+++ tg_sdk/policy/exceptions.py
@@ -1,5 +1,0 @@
-from tg_sdk.abstract import SDKException
-
-
-class NoBillingAddressException(SDKException):
-    message = "You must provide a billing address"