+++ ticketguardian/item/__init__.py
@@ -1,0 +1,3 @@
+from ticketguardian.item.item import Item
+
+__all__ = ['Item']